feat(articles detail): resolve two and three column images components, rename code highlighter component
