+++ src/components/ThreeColumnImages/styles.scss
@@ -1,0 +1,8 @@
+.three-column-images {
+  display: flex;
+}
+
+.three-column-image {
+  width: 300px;
+  padding: 4px 4px 4px 4px;
+}
