+++ src/components/TwoColumnImages/styles.scss
@@ -1,0 +1,9 @@
+.two-column-images {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.two-column-image {
+  width: 300px;
+  padding: 4px 4px 4px 4px;
+}
